Move SourceFileHeader to :build-logic:kotlin-dsl-shared-runtime
