+++ platforms/core-configuration/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/codegen/ApiExtensionsJar.kt
@@ -16,14 +16,13 @@
 
 package org.gradle.kotlin.dsl.codegen
 
-import org.gradle.kotlin.dsl.support.loggerFor
-import org.gradle.kotlin.dsl.support.compileToDirectory
-import org.gradle.kotlin.dsl.support.zipTo
-
 import org.gradle.api.internal.file.temp.TemporaryFileProvider
+import org.gradle.kotlin.dsl.internal.sharedruntime.codegen.kotlinDslPackagePath
 import org.gradle.kotlin.dsl.support.EmbeddedKotlinCompilerWarning
 import org.gradle.kotlin.dsl.support.bytecode.GradleJvmVersion
-
+import org.gradle.kotlin.dsl.support.compileToDirectory
+import org.gradle.kotlin.dsl.support.loggerFor
+import org.gradle.kotlin.dsl.support.zipTo
 import java.io.File
